Rename flaskext as flaskutils and move it to core
@maxfliri
