+++ backdrop/write/api.py
@@ -1,18 +1,16 @@
 from os import getenv
 
-from flask import Flask, request, jsonify, g, abort
+from flask import Flask, request, jsonify, g
 from backdrop import statsd
 from backdrop.core.log_handler \
     import create_request_logger, create_response_logger
-from backdrop.write.flaskext import BucketConverter
+from backdrop.core.flaskutils import BucketConverter
 from backdrop.write.permissions import Permissions
 from backdrop.write.admin_ui import use_single_sign_on
 from backdrop.write import admin_ui, parse_and_store
 
 from ..core.errors import ParseError, ValidationError
-from ..core.validation import bucket_is_valid
-from ..core import database, log_handler, records, cache_control
-from ..core.bucket import Bucket
+from ..core import database, log_handler, cache_control
 
 from .validation import bearer_token_is_valid
